requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

beautifulsoup4==4.11.2
mysql-connector-python==8.0.33
pandas==1.5.3
python-dotenv==1.0.0
requests==2.28.2
scraper==0.1.0
